Move handler to seperate files
